grub-efi: switch rollback item to backup partition
This commit introduces a new grub env var - rollback_part, which
indicates the backup partition.

The default value of this var is "_b", which means otaboot_b|otaroot_b
is backup partitions. After the system upgrade, this value should be
cleaned if it is "_b" currently, or set it to "_b" if it is empty.

So rollback_part should be switched over between "_b" and "". If regular
boot|root partition is otaboot|otaroot, this value is "_b", otherwise,
if regular boot|root partition is otaboot_b|otaroot_b, this value is
"empty".

And system upgrade manager should take care of the manipulation of this
var.

Signed-off-by: Yunguo Wei <yunguo.wei@windriver.com>