systemd:Mount /var in journald service
When the system reboots, a failed message will be shown on unmounting /var.

The root cause is that systemd-journald is still running when umounting /var,
and systemd-journald is outputting log to /var/log/journal.

So we add a dependency definition to make sure /var is umounted after
systemd--journald is stopped.

Signed-off-by: Jun Miao <jun.miao@windriver.com>
[ywei: Rephrase long git log]
Signed-off-by: Yunguo Wei <yunguo.wei@windriver.com>